Pandoc conversion of Tutorial chapters into RST
(A) Initial approach:

$ for f in *.tex; do pandoc -f latex -t rst -o ../${f%.*}.rst $f; done

This does not handle the references (e.g. to chapters) properly.

(B) However, this does - it just needs cutting up again:

$ pandoc -f latex -t rst -o Tutorial.rst Tutorial.tex

Wrote a simple script to do this (after earlier standardising the
chapter labels to help):

$ pandoc -f latex -t rst Tutorial.tex | ../.github/rst_chunk.py

Fixes the chapter numbers, but dumps the references to the
end of the appendix (poorly formatted anyway).

This then fixes links to other chapters (since pandoc was writing
a single large RST file, it made all the links internal HTML style
starting with hash):

$ sed -i.bak "s/<#chapter:/<chapter:/g" chapter_*.rst

(C) Wrote a script to convert the links to :ref: style instead,
First, hack the comment lines which otherwise pandoc ignores,
the convert to RST:

$ sed -i.bak "s#^%doctest#\\\pandocignoreme .. doctest#g" *.tex
$ sed -i.bak "s#^%cont-doctest#\\\pandocignoreme .. cont-doctest#g" *.tex
$ for f in *.tex; do pandoc -f latex -t rst -o ../${f%.*}.rst $f; done
$ git checkout *.tex

Then fix links:

$ ../.github/ref_fix.py chapter_*.rst